+++ models/gif_psc_exp_multisynapse.h
@@ -420,7 +420,7 @@ gif_psc_exp_multisynapse::send_test_event( Node& target, size_t receptor_type, s
 inline size_t
 gif_psc_exp_multisynapse::handles_test_event( SpikeEvent&, size_t receptor_type )
 {
-  if ( receptor_type <= 0 or receptor_type > static_cast< size_t >( P_.n_receptors_() ) )
+  if ( receptor_type <= 0 or receptor_type > P_.n_receptors_() )
   {
     throw IncompatibleReceptorType( receptor_type, get_name(), "SpikeEvent" );
   }
