+++ models/glif_cond.cpp
@@ -767,7 +767,7 @@ nest::glif_cond::update( Time const& origin, const long from, const long to )
 size_t
 nest::glif_cond::handles_test_event( SpikeEvent&, size_t receptor_type )
 {
-  if ( receptor_type <= 0 or receptor_type > static_cast< size_t >( P_.n_receptors_() ) )
+  if ( receptor_type <= 0 or receptor_type > P_.n_receptors_() )
   {
     throw IncompatibleReceptorType( receptor_type, get_name(), "SpikeEvent" );
   }
