+++ models/iaf_psc_exp_multisynapse.cpp
@@ -347,7 +347,7 @@ iaf_psc_exp_multisynapse::update( const Time& origin, const long from, const lon
 size_t
 iaf_psc_exp_multisynapse::handles_test_event( SpikeEvent&, size_t receptor_type )
 {
-  if ( receptor_type <= 0 or receptor_type > static_cast< size_t >( P_.n_receptors_() ) )
+  if ( receptor_type <= 0 or receptor_type > P_.n_receptors_() )
   {
     throw IncompatibleReceptorType( receptor_type, get_name(), "SpikeEvent" );
   }
